feat: move utils to utils package and make @excalidraw/utils a workspace
