+++ packages/excalidraw/global.d.ts
@@ -116,3 +116,5 @@ declare namespace jest {
     toBeNonNaNNumber(): void;
   }
 }
+
+declare module "@excalidraw/utils";
