+++ packages/excalidraw/package.json
@@ -1,6 +1,8 @@
 {
   "name": "@excalidraw/excalidraw",
   "version": "0.17.1",
+  "type": "module",
+  "module": "./dist/esm/index.js",
   "main": "main.js",
   "types": "types/packages/excalidraw/index.d.ts",
   "files": [
@@ -48,10 +50,12 @@
     "@excalidraw/laser-pointer": "1.2.0",
     "@excalidraw/mermaid-to-excalidraw": "0.2.0",
     "@excalidraw/random-username": "1.1.0",
+    "@excalidraw/utils": "workspace",
     "@radix-ui/react-popover": "1.0.3",
     "@radix-ui/react-tabs": "1.0.2",
     "@testing-library/jest-dom": "5.16.2",
     "@testing-library/react": "12.1.5",
+    "@tldraw/tldraw": "1.29.2",
     "@tldraw/vec": "1.7.1",
     "browser-fs-access": "0.29.1",
     "canvas-roundrect-polyfill": "0.0.1",
@@ -77,9 +81,6 @@
     "tunnel-rat": "0.1.2"
   },
   "devDependencies": {
-    "@types/pako": "1.0.3",
-    "@types/pica": "5.1.3",
-    "@types/resize-observer-browser": "0.1.7",
     "@babel/core": "7.18.9",
     "@babel/plugin-transform-arrow-functions": "7.18.6",
     "@babel/plugin-transform-async-to-generator": "7.18.6",
@@ -89,6 +90,9 @@
     "@babel/preset-react": "7.18.6",
     "@babel/preset-typescript": "7.18.6",
     "@size-limit/preset-big-lib": "9.0.0",
+    "@types/pako": "1.0.3",
+    "@types/pica": "5.1.3",
+    "@types/resize-observer-browser": "0.1.7",
     "autoprefixer": "10.4.7",
     "babel-loader": "8.2.5",
     "babel-plugin-transform-class-properties": "6.24.1",
@@ -116,6 +120,7 @@
     "gen:types": "tsc --project tsconfig-types.json",
     "build:umd": "rm -rf dist && cross-env NODE_ENV=production webpack --config webpack.prod.config.js && cross-env NODE_ENV=development webpack --config webpack.dev.config.js && NODE_ENV=development webpack --config webpack.preact.config.js && NODE_ENV=production webpack --config webpack.preact.config.js && yarn gen:types",
     "build:umd:withAnalyzer": "cross-env NODE_ENV=production ANALYZER=true webpack --config webpack.prod.config.js",
+    "build:esm": "tsc",
     "pack": "yarn build:umd && yarn pack",
     "start": "webpack serve --config webpack.dev-server.config.js",
     "build:example": "EXAMPLE=true webpack --config webpack.dev-server.config.js && yarn gen:types",
