+++ packages/utils/export.ts
@@ -1,23 +1,23 @@
 import {
   exportToCanvas as _exportToCanvas,
   exportToSvg as _exportToSvg,
-} from "./excalidraw/scene/export";
-import { getDefaultAppState } from "./excalidraw/appState";
-import { AppState, BinaryFiles } from "./excalidraw/types";
+} from "../excalidraw/scene/export";
+import { getDefaultAppState } from "../excalidraw/appState";
+import { AppState, BinaryFiles } from "../excalidraw/types";
 import {
   ExcalidrawElement,
   ExcalidrawFrameLikeElement,
   NonDeleted,
-} from "./excalidraw/element/types";
-import { restore } from "./excalidraw/data/restore";
-import { MIME_TYPES } from "./excalidraw/constants";
-import { encodePngMetadata } from "./excalidraw/data/image";
-import { serializeAsJSON } from "./excalidraw/data/json";
+} from "../excalidraw/element/types";
+import { restore } from "../excalidraw/data/restore";
+import { MIME_TYPES } from "../excalidraw/constants";
+import { encodePngMetadata } from "../excalidraw/data/image";
+import { serializeAsJSON } from "../excalidraw/data/json";
 import {
   copyBlobToClipboardAsPng,
   copyTextToSystemClipboard,
   copyToClipboard,
-} from "./excalidraw/clipboard";
+} from "../excalidraw/clipboard";
 
 export { MIME_TYPES };
 
@@ -215,11 +215,11 @@ export {
 export {
   serializeAsJSON,
   serializeLibraryAsJSON,
-} from "./excalidraw/data/json";
+} from "../excalidraw/data/json";
 export {
   loadFromBlob,
   loadSceneOrLibraryFromBlob,
   loadLibraryFromBlob,
-} from "./excalidraw/data/blob";
-export { getFreeDrawSvgPath } from "./excalidraw/renderer/renderElement";
-export { mergeLibraryItems } from "./excalidraw/data/library";
+} from "../excalidraw/data/blob";
+export { getFreeDrawSvgPath } from "../excalidraw/renderer/renderElement";
+export { mergeLibraryItems } from "../excalidraw/data/library";
